rtl8139: disable pxe rom when xen is enabled
When QEMU allocate RAM address it's always starting with the same base address.
But, in case of QEMU disaggregation if the VGA is emulated in another QEMU than
net, they will both have the same ram address.

It's an ugly fix, but I don't find the right way to modify it.